@@ -6,19 +6,22 @@
     "start": "node node_modules/react-native/local-cli/cli.js start",
     "test": "jest",
     "bundle_uwp": "node node_modules/react-native/local-cli/cli.js bundle --dev false --platform windows --entry-file index.windows.js --bundle-output windows/taa/ReactAssets/index.windows.bundle --assets-dest windows/taa/ReactAssets",
-    "cps_uwp": "code-push release-react taa-uwp windows",
-    "cpp_uwp": "code-push promote taa-uwp Staging Production",
-    "cps_and": "code-push release-react taa-android android",
-    "cpp_and": "code-push promote taa-android Staging Production",
+    "cps_uwp": "appcenter codepush release-react -a myh/taa-uwp",
+    "cpp_uwp": "appcenter codepush promote -a myh/taa-uwp -s Staging -d Production",
+    "cps_and": "appcenter codepush release-react -a myh/taa-and",
     "build_apk": "cd android && gradlew assembleRelease",
+    "cpp_and": "appcenter codepush promote -a myh/taa-and -s Staging -d Production",
     "install_apk": "adb install ./android/app/build/outputs/apk/app-release.apk"
   },
   "dependencies": {
+    "@babel/plugin-proposal-decorators": "^7.2.3",
     "axios": "^0.16.1",
-    "react": "16.0.0-alpha.6",
-    "react-native": "^0.43.4",
-    "react-native-code-push": "^2.0.2-beta",
-    "react-native-windows": "^0.43.0-rc.0",
+    "create-react-class": "^15.6.3",
+    "react": "16.6.3",
+    "react-native": "0.57.8",
+    "react-native-code-push": "2.1.0-beta",
+    "react-native-windows": "^0.57.0-rc.0",
+    "react-navigation": "^1.6.1",
     "react-redux": "^5.0.4",
     "react-thunk": "^1.0.0",
     "redux": "^3.6.0",
@@ -27,13 +30,14 @@
     "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "babel-jest": "19.0.0",
+    "babel-jest": "23.6.0",
     "babel-plugin-transform-decorators": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-react-native": "1.9.1",
-    "jest": "19.0.2",
-    "react-test-renderer": "15.4.2",
-    "rnpm-plugin-windows": "^0.2.4",
+    "jest": "23.6.0",
+    "metro-react-native-babel-preset": "0.51.1",
+    "react-test-renderer": "16.6.3",
+    "rnpm-plugin-windows": "^0.2.8",
     "tslint": "^5.1.0",
     "typescript": "^2.2.2",
     "typings": "^2.1.1"
