@@ -1,21 +1,13 @@
 using ReactNative;
-using ReactNative.Modules.Launch;
-using System;
-using Windows.ApplicationModel;
-using Windows.ApplicationModel.Activation;
-using Windows.UI.Core;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Navigation;
 
 namespace taa
 {
     /// <summary>
     /// Provides application-specific behavior to supplement the default Application class.
     /// </summary>
-    sealed partial class App : Application
+    sealed partial class App : ReactApplication
     {
-        private readonly ReactPage _reactPage;
+        private readonly ReactNativeHost _host = new MainReactNativeHost();
 
         /// <summary>
         /// Initializes the singleton application object.  This is the first line of authored code
@@ -24,124 +16,11 @@ namespace taa
         public App()
         {
             this.InitializeComponent();
-            this.Suspending += OnSuspending;
-            this.Resuming += OnResuming;
-
-            _reactPage = new MainPage();
         }
 
         /// <summary>
-        /// Invoked when the application is launched normally by the end user.  Other entry points
-        /// will be used such as when the application is launched to open a specific file.
+        /// The React Native host.
         /// </summary>
-        /// <param name="e">Details about the launch request and process.</param>
-        protected override void OnLaunched(LaunchActivatedEventArgs e)
-        {
-            base.OnLaunched(e);
-            OnCreate(e.Arguments);
-        }
-
-        /// <summary>
-        /// Invoked when the application is activated.
-        /// </summary>
-        /// <param name="args">The activated event arguments.</param>
-        protected override void OnActivated(IActivatedEventArgs args)
-        {
-            base.OnActivated(args);
-
-            switch (args.Kind)
-            {
-                case ActivationKind.Protocol:
-                case ActivationKind.ProtocolForResults:
-                    var protocolArgs = (IProtocolActivatedEventArgs)args;
-                    LauncherModule.SetActivatedUrl(protocolArgs.Uri.AbsoluteUri);
-                    break;
-            }
-
-            if (args.PreviousExecutionState != ApplicationExecutionState.Running &&
-                args.PreviousExecutionState != ApplicationExecutionState.Suspended)
-            {
-                OnCreate(null);
-            }
-        }
-
-        /// <summary>
-        /// Called whenever the app is opened to initia
-        /// </summary>
-        /// <param name="arguments"></param>
-        private void OnCreate(string arguments)
-        {
-            _reactPage.OnResume(Exit);
-
-#if DEBUG
-            if (System.Diagnostics.Debugger.IsAttached)
-            {
-                this.DebugSettings.EnableFrameRateCounter = true;
-            }
-
-            SystemNavigationManager.GetForCurrentView().AppViewBackButtonVisibility =
-                AppViewBackButtonVisibility.Visible;
-#endif
-
-            Frame rootFrame = Window.Current.Content as Frame;
-
-            // Do not repeat app initialization when the Window already has content,
-            // just ensure that the window is active
-            if (rootFrame == null)
-            {
-                _reactPage.OnCreate(arguments);
-
-                // Create a Frame to act as the navigation context and navigate to the first page
-                rootFrame = new Frame();
-
-                rootFrame.NavigationFailed += OnNavigationFailed;
-
-                // Place the frame in the current Window
-                Window.Current.Content = rootFrame;
-            }
-
-            if (rootFrame.Content == null)
-            {
-                // When the navigation stack isn't restored navigate to the first page,
-                // configuring the new page by passing required information as a navigation
-                // parameter
-                rootFrame.Content = _reactPage;
-            }
-
-            // Ensure the current window is active
-            Window.Current.Activate();
-        }
-
-        /// <summary>
-        /// Invoked when Navigation to a certain page fails
-        /// </summary>
-        /// <param name="sender">The Frame which failed navigation</param>
-        /// <param name="e">Details about the navigation failure</param>
-        private void OnNavigationFailed(object sender, NavigationFailedEventArgs e)
-        {
-            throw new Exception("Failed to load Page " + e.SourcePageType.FullName);
-        }
-
-        /// <summary>
-        /// Invoked when application execution is being suspended.  Application state is saved
-        /// without knowing whether the application will be terminated or resumed with the contents
-        /// of memory still intact.
-        /// </summary>
-        /// <param name="sender">The source of the suspend request.</param>
-        /// <param name="e">Details about the suspend request.</param>
-        private void OnSuspending(object sender, SuspendingEventArgs e)
-        {
-            _reactPage.OnSuspend();
-        }
-
-        /// <summary>
-        /// Invoked when application execution is being resumed.
-        /// </summary>
-        /// <param name="sender">The source of the resume request.</param>
-        /// <param name="e">Details about the resume request.</param>
-        private void OnResuming(object sender, object e)
-        {
-            _reactPage.OnResume(Exit);
-        }
+        public override ReactNativeHost Host => _host;
     }
 }
